moved the location of the form to the story show page, added a link to the story show page that directs you to the bottom of the page. Added a redirect to The event show page on form submission
