+++ app/javascript/react/components/StoryShowContainer.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import EventTile from './EventTile';
+import PossibilityFormContainer from './PossibilityFormContainer'
 
 class StoryShowContainer extends Component {
   constructor(props) {
@@ -8,6 +9,7 @@ class StoryShowContainer extends Component {
       story: {},
       events: []
     }
+    this.addNewPossibility = this.addNewPossibility.bind(this)
   }
 
   componentDidMount() {
@@ -29,6 +31,33 @@ class StoryShowContainer extends Component {
           story: story,
           events: story.events
         } )
+      })
+      .catch(error => console.error(`Error in fetch: ${error.message}`));
+  }
+
+  addNewPossibility(formPayload) {
+    const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content')
+    fetch('/api/v1/possibilities', {
+      method: 'POST',
+      body: JSON.stringify(formPayload),
+      headers: {
+        'Content-Type': 'application/json',
+        'X-CSRF-Token': csrfToken
+      },
+      credentials: 'same-origin'
+    })
+      .then(response => {
+        if (response.ok) {
+          return response;
+        } else {
+          let errorMessage = `${response.status} (${response.statusText})`,
+           error = new Error(errorMessage);
+          throw(error);
+        }
+      })
+      .then(response => response.json())
+      .then(body => {
+        return window.location.href = `/stories/${this.state.story.id}/events/${this.state.events.slice(-1)[0].id}`
       })
       .catch(error => console.error(`Error in fetch: ${error.message}`));
   }
@@ -55,11 +84,22 @@ class StoryShowContainer extends Component {
         )
       }
     })
+    let lastEvent
+    if (this.state.events.length > 0) {
+      lastEvent = this.state.events.slice(-1)[0].id
+    }
     return(
       <div>
         <h1>{ this.state.story.name }</h1>
         <p>{ this.state.story.description }</p>
+        <a href="#form">Click Here To Jump To The Bottom Of The Page</a>
         <ul>{ events }</ul>
+        <div id="form">
+          <PossibilityFormContainer
+            addNewPossibility={this.addNewPossibility}
+            eventID={lastEvent}
+          />
+        </div>
       </div>
     )
   }
