+++ app/serializers/story_serializer.rb
@@ -1,4 +1,9 @@
 class StorySerializer < ActiveModel::Serializer
-  attributes :id, :name, :description, :events
+  attributes :id, :name, :description, :events, :user_id
   has_many :events
+  def user_id
+    if current_user
+      current_user.id
+    end
+  end
 end
